pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "clawarm-bridge"
version = "0.1.0"
description = "Python bridge server for AI-powered robotic arm control via OpenClaw"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Clawland AI", email = "hello@clawland.ai" }]
keywords = ["robotic-arm", "openclaw", "pyagxarm", "nero", "piper", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-can>=4.4.0",
]

[project.optional-dependencies]
arm = ["pyAgxArm"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/Clawland-AI/clawarm"
Repository = "https://github.com/Clawland-AI/clawarm"
Issues = "https://github.com/Clawland-AI/clawarm/issues"

[project.scripts]
clawarm-bridge = "bridge.server:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
